pyproject.toml
[tool.poetry]
name = "hitakort"
version = "0.1.0"
description = "A Telegram bot for tracking and visualizing grid-based hit data."
authors = ["Nachtalb <na@nachtalb.io>"]
readme = "README.md"
license = "LGPL-3.0"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
numpy = "^1.26.4"
python-telegram-bot = { extras = ["ext", "webhooks"], version = "^21.2" }

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
ipdb = "^0.13.13"
mypy = "^1.10.0"
isort = "^5.13.2"
pre-commit = "^3.7.1"
ruff-lsp = "^0.0.53"
types-pillow = "^10.2.0.20240520"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
line_length = 120

[tool.black]
line-length = 120
preview = true
target-version = ["py312"]

[tool.ruff]
line-length = 120

[tool.mypy]
python_version = "3.12"
show_error_codes = true
pretty = true
strict = true

[[tool.mypy.overrides]]
module = 'tests'
no_warn_no_return = true

[tool.poetry.scripts]
hitakort = "hitakort.bot:main"
